Added business logic for Orders, some other changes and views.
Bug fixes and refactoring after review.

+++ src/main/java/com/ivan/salesapp/services/CategoryService.java
@@ -10,6 +10,8 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import static java.util.stream.Collectors.toList;
+
 @Service
 public class CategoryService implements ICategoryService {
     private final CategoryRepository categoryRepository;
@@ -31,7 +33,7 @@ public class CategoryService implements ICategoryService {
     @Override
     public List<CategoryServiceModel> findAllCategories() {
         return this.categoryRepository.findAll().stream()
-                .map(c -> this.modelMapper.map(c, CategoryServiceModel.class)).collect(Collectors.toList());
+                .map(c -> this.modelMapper.map(c, CategoryServiceModel.class)).collect(toList());
     }
 
     @Override
@@ -54,9 +56,15 @@ public class CategoryService implements ICategoryService {
     }
 
     @Override
-    public void deleteCategory(String id) {
+    public void deleteCategory(String id, IProductService iProductService) {
         Category category = this.categoryRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("Incorrect id!"));
+
+        int allProductsByCategory = iProductService.findAllByCategory(category.getName()).size();
+
+        if (allProductsByCategory != 0){
+            throw new IllegalArgumentException("There are no products in this category!");
+        }
 
         this.categoryRepository.delete(category);
     }
